cleanup

+++ include/steepest_descent/gauss_laguerre.h
@@ -8,7 +8,7 @@ namespace gauss_laguerre {
 
 
 
-	auto calculate_laguerre_points_and_weights(int n) -> std::tuple<std::vector<double>, std::vector<double>>;
+	auto calculate_laguerre_points_and_weights(size_t n) -> std::tuple<std::vector<double>, std::vector<double>>;
 	
 	auto calculate_integral_cauchy(const path_utils::path_function paths, std::vector<double> nodes, std::vector<double> weights)->std::complex<double>;
